Change boot.h so that it resets the hw bank id at the start of every test for qsr

+++ tests/python_tests/test_pack_quasar.py
@@ -27,8 +27,8 @@ from helpers.utils import passed_test
     formats=input_output_formats(
         [
             DataFormat.Float16_b,
-            # DataFormat.Float16,
-            DataFormat.Float32,
+            DataFormat.Float16,
+            # DataFormat.Float32,
         ],
         same=True,
     ),
