+++ tests/python_tests/test_unpack_unary_operand_quasar.py
@@ -65,8 +65,8 @@ def generate_unpack_unary_operand_combinations(
 UNPACK_FORMATS = input_output_formats(
     [
         DataFormat.Float16_b,
-        # DataFormat.Float16,
-        DataFormat.Float32,
+        DataFormat.Float16,
+        # DataFormat.Float32,
     ]
 )
 ALL_UNPACK_UNARY_OPERAND_COMBINATIONS = generate_unpack_unary_operand_combinations(
